+++ build.zig
@@ -10,7 +10,8 @@ const std = @import("std");
 const builtin = @import("builtin");
 const log = std.log.scoped(.zdt_build);
 
-const zdt_version = std.SemanticVersion{ .major = 0, .minor = 6, .patch = 2 };
+const zdt_version = std.SemanticVersion{ .major = 0, .minor = 6, .patch = 3 };
+
 const tzdb_tag = "2025b";
 
 const example_files = [_][]const u8{