+++ lib/Timezone.zig
@@ -1,6 +1,7 @@
 //! A set of rules to describe date and time somewhere on earth, relative to universal time (UTC).
 
 const std = @import("std");
+const assert = std.debug.assert;
 const builtin = @import("builtin");
 const log = std.log.scoped(.zdt__Timezone);
 
@@ -15,6 +16,7 @@ const Timezone = @This();
 
 /// embedded IANA time zone database (eggert/tz)
 pub const tzdata = @import("./tzdata.zig").tzdata;
+pub const sizeOftzdata = @import("./tzdata.zig").sizeOftzdata;
 
 pub const tzdb_version = @import("./tzdata.zig").tzdb_version;
 
@@ -22,24 +24,24 @@ pub const tzdb_version = @import("./tzdata.zig").tzdb_version;
 // anonymous import; see build.zig
 pub const tzdb_prefix = @import("tzdb_prefix").tzdb_prefix;
 
-/// Where to comptime-load IANA tz database files from
-const comptime_tzdb_prefix = "./tzdata/zoneinfo/"; // IANA db as provided by the library
-
 // longest tz name is 'America/Argentina/ComodRivadavia' --> 32 ASCII chars
 const cap_name_data: usize = 32;
 
 const ruleTypes = enum {
     tzif,
+    tzif_fixedsize,
     posixtz,
     utc,
 };
 
 /// rule sources for a time zone
 rules: union(ruleTypes) {
-    /// IANA tz-db/tzdata TZif file;
+    /// IANA tzdb/tzdata TZif file;
     /// use Timezone.fromTzdata or Timezone.fromSystemTzdata to set as time zone of a datetime.
-    tzif: tzif.Tz,
-    /// Not implemented! - POSIX TZ string
+    tzif: tzif.TzAlloc,
+    /// IANA tzdb/tzdata TZif file in a fixed-size data structure - no allocator required!
+    tzif_fixedsize: tzif.Tz,
+    /// POSIX TZ string
     posixtz: psx.PosixTz,
     /// UTC placeholder;
     /// use Timezone.UTC constant to set UTC as time zone of a datetime.
@@ -66,6 +68,7 @@ pub fn name(tz: *const Timezone) []const u8 {
 pub fn fromPosixTz(posixString: []const u8) !Timezone {
     const ptz = try psx.parsePosixTzString(posixString);
     var tz = Timezone{ .rules = .{ .posixtz = ptz } };
+    assert(posixString.len <= cap_name_data);
     tz.__name_data_len = if (posixString.len <= cap_name_data) posixString.len else cap_name_data;
     @memcpy(tz.__name_data[0..tz.__name_data_len], posixString[0..tz.__name_data_len]);
     return tz;
@@ -74,20 +77,27 @@ pub fn fromPosixTz(posixString: []const u8) !Timezone {
 /// Make a time zone from IANA tz database TZif data, taken from the embedded tzdata.
 /// The caller must make sure to de-allocate memory used for storing the TZif file's content
 /// by calling the deinit method of the returned TZ instance.
-pub fn fromTzdata(identifier: []const u8, allocator: std.mem.Allocator) TzError!Timezone {
+///
+/// Note that the allocator is optional. If 'null' is provided instead of an allocator,
+/// a fixed-size structure will be used to holde the timezone rules, instead of doing this
+/// dynamically in heap memory. This is faster, but requires more memory overall.
+pub fn fromTzdata(identifier: []const u8, allocator: ?std.mem.Allocator) TzError!Timezone {
     if (!identifierValid(identifier)) return TzError.InvalidIdentifier;
 
     if (std.mem.eql(u8, identifier, "localtime")) return tzLocal(allocator);
 
     if (tzdata.get(identifier)) |TZifBytes| {
         var in_stream = std.io.fixedBufferStream(TZifBytes);
-        const tzif_tz = tzif.Tz.parse(allocator, in_stream.reader()) catch
-            return TzError.TZifUnreadable;
+        var tz = Timezone{ .rules = .{ .utc = .{} } };
 
-        // ensure that there is a footer: requires v2+ TZif files.
-        _ = tzif_tz.footer orelse return TzError.BadTZifVersion;
+        if (allocator) |alcr| {
+            const tzif_tz = tzif.TzAlloc.parse(alcr, in_stream.reader()) catch return TzError.TZifUnreadable;
+            tz = Timezone{ .rules = .{ .tzif = tzif_tz } };
+        } else {
+            const tzif_tz = tzif.Tz.parse(in_stream.reader()) catch return TzError.TZifUnreadable;
+            tz = Timezone{ .rules = .{ .tzif_fixedsize = tzif_tz } };
+        }
 
-        var tz = Timezone{ .rules = .{ .tzif = tzif_tz } };
         tz.__name_data_len = if (identifier.len <= cap_name_data) identifier.len else cap_name_data;
         @memcpy(tz.__name_data[0..tz.__name_data_len], identifier[0..tz.__name_data_len]);
 
@@ -101,7 +111,11 @@ pub fn fromTzdata(identifier: []const u8, allocator: std.mem.Allocator) TzError!
 /// To use the system's tzdata, use 'zdt.Timezone.tzdb_prefix'.
 /// The caller must make sure to de-allocate memory used for storing the TZif file's content
 /// by calling the deinit method of the returned Timezone instance.
-pub fn fromSystemTzdata(identifier: []const u8, db_path: []const u8, allocator: std.mem.Allocator) TzError!Timezone {
+///
+/// Note that the allocator is optional. If 'null' is provided instead of an allocator,
+/// a fixed-size structure will be used to holde the timezone rules, instead of doing this
+/// dynamically in heap memory. This is faster, but requires more memory overall.
+pub fn fromSystemTzdata(identifier: []const u8, db_path: []const u8, allocator: ?std.mem.Allocator) TzError!Timezone {
     if (!identifierValid(identifier)) return TzError.InvalidIdentifier;
     var path_buffer: [std.fs.max_path_bytes]u8 = undefined;
     var fba = std.heap.FixedBufferAllocator.init(&path_buffer);
@@ -113,13 +127,16 @@ pub fn fromSystemTzdata(identifier: []const u8, db_path: []const u8, allocator: 
     const file = std.fs.openFileAbsolute(p, .{}) catch return TzError.TZifUnreadable;
     defer file.close();
 
-    const tzif_tz = tzif.Tz.parse(allocator, file.reader()) catch
-        return TzError.TZifUnreadable;
+    var tz = Timezone{ .rules = .{ .utc = .{} } };
 
-    // ensure that there is a footer: requires v2+ TZif files.
-    _ = tzif_tz.footer orelse return TzError.BadTZifVersion;
+    if (allocator) |alcr| {
+        const tzif_tz = tzif.TzAlloc.parse(alcr, file.reader()) catch return TzError.TZifUnreadable;
+        tz = Timezone{ .rules = .{ .tzif = tzif_tz } };
+    } else {
+        const tzif_tz = tzif.Tz.parse(file.reader()) catch return TzError.TZifUnreadable;
+        tz = Timezone{ .rules = .{ .tzif_fixedsize = tzif_tz } };
+    }
 
-    var tz = Timezone{ .rules = .{ .tzif = tzif_tz } };
     // default: use identifier as name
     tz.__name_data_len = if (identifier.len <= cap_name_data) identifier.len else cap_name_data;
     @memcpy(tz.__name_data[0..tz.__name_data_len], identifier[0..tz.__name_data_len]);
@@ -131,6 +148,7 @@ pub fn fromSystemTzdata(identifier: []const u8, db_path: []const u8, allocator: 
         const part = pathname_iterator.next() orelse identifier;
         if (!std.mem.eql(u8, identifier, part)) {
             const tmp_name = pathname_iterator.next() orelse "?";
+            assert(tmp_name.len <= cap_name_data);
             // we might need to overwrite pre-defined data with zeros:
             var name_data = std.mem.zeroes([cap_name_data]u8);
             const len: usize = if (tmp_name.len <= cap_name_data) tmp_name.len else cap_name_data;
@@ -150,6 +168,7 @@ pub fn deinit(tz: *Timezone) void {
 
     switch (tz.rules) {
         .tzif => |*_tzif| _tzif.deinit(),
+        .tzif_fixedsize => return,
         .posixtz => return,
         .utc => return,
     }
@@ -160,7 +179,7 @@ pub fn deinit(tz: *Timezone) void {
 /// Note: Windows does not use the IANA time zone database;
 /// a mapping from Windows db to IANA db is prone to errors.
 /// Use with caution.
-pub fn tzLocal(allocator: std.mem.Allocator) TzError!Timezone {
+pub fn tzLocal(allocator: ?std.mem.Allocator) TzError!Timezone {
     switch (builtin.os.tag) {
         .linux, .macos => {
             const default_path = "/etc/localtime";
@@ -186,13 +205,19 @@ pub fn format(
 ) !void {
     _ = fmt;
     _ = options;
-    try writer.print("Time zone, name: {c}", .{tz.name()});
+    try writer.print("Time zone, name: {s}. ", .{tz.name()});
+    switch (tz.rules) {
+        .tzif => try writer.print("Source: TZif. Memory: dynamic.", .{}),
+        .tzif_fixedsize => try writer.print("Source: TZif. Memory: static.", .{}),
+        .posixtz => try writer.print("Source: POSIX string.", .{}),
+        else => return,
+    }
 }
 
 /// Time zone identifiers must only contain alpha-numeric characters
 /// as well as '+', '-', '_' and '/' (path separator).
-pub fn identifierValid(idf: []const u8) bool {
-    for (idf, 0..) |c, i| {
+pub fn identifierValid(id: []const u8) bool {
+    for (id, 0..) |c, i| {
         switch (c) {
             // OK cases:
             'A'...'Z',
@@ -207,8 +232,8 @@ pub fn identifierValid(idf: []const u8) bool {
             // - not last char
             // - next char is not a period
             '.' => {
-                if (i == idf.len - 1) return false;
-                if (idf[i + 1] == '.') return false;
+                if (i == id.len - 1) return false;
+                if (id[i + 1] == '.') return false;
                 continue;
             },
             else => return false,
@@ -218,7 +243,7 @@ pub fn identifierValid(idf: []const u8) bool {
 }
 
 test "embed TZif from lib dir" {
-    const tzfile = comptime_tzdb_prefix ++ "Europe/Berlin";
+    const tzfile = "./tzdata/zoneinfo/Europe/Berlin";
     const data = @embedFile(tzfile);
     var in_stream = std.io.fixedBufferStream(data);
     var tz = try std.Tz.parse(std.testing.allocator, in_stream.reader());